move: methods about notion into another dir
